add update method and transfer operation
add update method in dao, which sets new values in database
add transfer operation to move money from one account to another

+++ src/by/epam/kisel/util/validation/Validator.java
@@ -48,11 +48,11 @@ public class Validator {
 				isNull(client.getRealAddress());
 	}
 	
-	public static boolean lessThanZero(int number) {
+	public static boolean lessThanZero(long number) {
 		return number < 0;
 	}
 	
-	public static boolean lessThanZeroOrEqualZero(int number) {
+	public static boolean lessThanZeroOrEqualZero(long number) {
 		return number <= 0;
 	}
